chore(loading): dedicated files for resources
